fix: move upload store (#3357)
