+++ pkg/localstorev2/internal/upload/uploadstore_test.go
@@ -14,12 +14,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ethersphere/bee/pkg/localstorev2/internal"
+	"github.com/ethersphere/bee/pkg/localstorev2/internal/upload"
 	chunktest "github.com/ethersphere/bee/pkg/storage/testing"
-	"github.com/ethersphere/bee/pkg/storagev2"
+	storage "github.com/ethersphere/bee/pkg/storagev2"
 	"github.com/ethersphere/bee/pkg/storagev2/inmemchunkstore"
 	inmem "github.com/ethersphere/bee/pkg/storagev2/inmemstore"
-	"github.com/ethersphere/bee/pkg/storagev2/internal"
-	"github.com/ethersphere/bee/pkg/storagev2/internal/upload"
 	"github.com/ethersphere/bee/pkg/storagev2/storagetest"
 	"github.com/ethersphere/bee/pkg/swarm"
 	swarmtesting "github.com/ethersphere/bee/pkg/swarm/test"
